+++ gemlib/a_form_dial.c
@@ -2,9 +2,9 @@
 #include "gem.h"
 
 
-int
-form_dial (int Flag, int Sx, int Sy, int Sw, int Sh, int Bx, int By, int Bw,
-	   int Bh)
+short
+form_dial (short Flag, short Sx, short Sy, short Sw, short Sh, short Bx, short By, short Bw,
+	   short Bh)
 {
 	aes_intin[0] = Flag;
 	aes_intin[1] = Sx;
@@ -15,11 +15,14 @@ form_dial (int Flag, int Sx, int Sy, int Sw, int Sh, int Bx, int By, int Bw,
 	aes_intin[6] = By;
 	aes_intin[7] = Bw;
 	aes_intin[8] = Bh;
+	
 	aes_control[0] = 51;
 	aes_control[1] = 9;
 	aes_control[2] = 1;
 	aes_control[3] = 0;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
