+++ gemlib/v_curleft.c
@@ -3,7 +3,7 @@
 
 
 void
-v_curleft (int handle)
+v_curleft (short handle)
 {
 	vdi_control[0] = 5;
 	vdi_control[1] = 0;
