+++ gemlib/v_curright.c
@@ -3,7 +3,7 @@
 
 
 void
-v_curright (int handle)
+v_curright (short handle)
 {
 	vdi_control[0] = 5;
 	vdi_control[1] = 0;
