+++ gemlib/v_pline.c
@@ -3,11 +3,11 @@
 
 
 void
-v_pline (int handle, int count, int pxyarray[])
+v_pline (short handle, short count, short pxyarray[])
 {
 	register short i;
 	register short *pts = &vdi_ptsin[0];
-	register int *pxy = &pxyarray[0];
+	register short *pxy = &pxyarray[0];
 
 	for (i = count; i != 0; i--)
 	{
